wip moving history state to shared module
