+++ src/app/shared/actions/history.actions.ts
@@ -4,7 +4,7 @@ import { HistoryEntity } from '../reducers/history.reducer';
 
 import {
     UpdateHistoryItemGamePayload, UpdateHistoryItemPlatformPayload, UpdateHistoryItemTimesPayload
-} from '../../../shared/models';
+} from '../models';
 
 export const ADD_NEW_HISTORY_ITEM = '[History] Add New History Item';
 export class AddNewHistoryItem implements Action {
