+++ src/app/shared/reducers/history.reducer.ts
